HACKATHON/weather-backend/requirements.txt
python-dotenv
requests
Flask
pandas
numpy
twilio
flask_cors
setuptools